# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking CRM — Mobile API",
        "description": "REST API для мобильного приложения мастера/сотрудника. Поддерживает офлайн-режим с дельта-синхронизацией. Авторизация через Bearer-токен (Sanctum).",
        "contact": {
            "name": "Booking CRM Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Production / Dev сервер"
        }
    ],
    "paths": {
...
        "/mobile/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Список клиентов",
                "description": "Поиск по имени, телефону, email. Поддерживает дельта-синхронизацию через updated_after.",
                "operationId": "7aeb7c01c7d4a9f55f8584a6a6a2f685",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по имени, телефону или email",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Анна"
                        }
                    },
                    {
                        "name": "updated_after",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-02-20T00:00:00Z"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список клиентов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "sync_timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-02-22T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Создать клиента",
                "operationId": "cba4e7e6a3e15830692d967d0aca8801",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Анна Петрова"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+79991234567",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "anna@example.com",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Предпочитает утренние записи",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Клиент создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Client"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации или лимит клиентов достигнут"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/mobile/clients/{id}": {
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Обновить клиента",
                "operationId": "cd1f3c6b3f25a9df225720d56cf31edd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Анна Петрова"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+79991234567",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Клиент обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Client"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Нет доступа"
                    },
                    "404": {
                        "description": "Клиент не найден"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/mobile/schedule": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Расписание мастера",
                "description": "Возвращает рабочее расписание по дням недели (7 записей). Время в timezone мастера.",
                "operationId": "3a54a90eb410f806920c577dc05ddace",
                "responses": {
                    "200": {
                        "description": "Расписание по дням недели",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "description": "Массив из 7 элементов (один на каждый день недели)",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Schedule"
                                            }
                                        },
                                        "sync_timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-02-22T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 50
                    },
                    "total": {
                        "type": "integer",
                        "example": 230
                    }
                },
                "type": "object"
            },
...
            "Client": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "name": {
                        "type": "string",
                        "example": "Анна Петрова"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+79991234567",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "example": "anna@example.com",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "example": "Предпочитает утренние записи",
                        "nullable": true
                    },
                    "total_visits": {
                        "type": "integer",
                        "example": 12
                    },
                    "total_spent": {
                        "type": "number",
                        "format": "float",
                        "example": 18500
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-20T08:30:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-01T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Schedule": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "day_of_week": {
                        "description": "0 = воскресенье, 1 = понедельник, ..., 6 = суббота (Carbon)",
                        "type": "integer",
                        "example": 1
                    },
                    "is_working": {
                        "type": "boolean",
                        "example": true
                    },
                    "start_time": {
                        "description": "Время начала работы HH:MM",
                        "type": "string",
                        "example": "09:00",
                        "nullable": true
                    },
                    "end_time": {
                        "description": "Время окончания работы HH:MM",
                        "type": "string",
                        "example": "18:00",
                        "nullable": true
                    },
                    "break_start": {
                        "description": "Начало перерыва HH:MM",
                        "type": "string",
                        "example": "13:00",
                        "nullable": true
                    },
                    "break_end": {
                        "description": "Конец перерыва HH:MM",
                        "type": "string",
                        "example": "14:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum",
                "description": "Введите токен, полученный из POST /api/mobile/login"
            }
        }
    },
...
}